requirements.txt
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
protobuf==3.20.3
numpy
pandas
matplotlib
opencv-python
